Bug 534874 - Fix preview server adaptors if java ee implementations have unexpected BSNs
To allow downstream integrators to use any implementation of the
java ee APIs, we should avoid hard-coding bundle names for those
APIs. This change makes the preview server adapters ask the framework
for the bundle names instead, based on class names that we know
must be present in such bundles.

Change-Id: Iad600de7f795ebcc71603fae82b2ceb22fe1f9c6
Signed-off-by: Mat Booth <mat.booth@redhat.com>